clean up model_data.csv reads
a set of function was made to serve this purpose, however they were never used. it seems to make more sense to just use imports. so did that.  this insures the file is read in only one part of the code and generally deduplicates the code which should make it easier to maintain